requirements.txt
# Core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1

# Production WSGI server (also used inside Docker)
gunicorn==21.2.0

# PostgreSQL driver
# Installs fine on Linux / inside Docker (pre-built wheel).
# Windows users without PostgreSQL headers: comment this out and use SQLite locally.
psycopg2-binary==2.9.9

# Testing
pytest==8.1.1
pytest-flask==1.3.0

# Linting
flake8==7.0.0
# psycopg2-binary==2.9.9
